Fix standards tests
