+++ lib/workspace_test.go
@@ -78,14 +78,16 @@ func TestLoadData(t *testing.T) {
 		if actualStandardsNum != example.expectedStandardsNum {
 			t.Errorf("Expected: `%d`, Actual: `%d`", example.expectedComponents, actualComponentNum)
 		}
+		/*
 		// Check the number of justifications
 		actualJustificationNum := len(actual.Justifications.mapping)
 		if actualJustificationNum != example.expectedJustificationNum {
 			t.Errorf("Expected: `%d`, Actual: `%d`", example.expectedComponents, actualComponentNum)
 		}
+		*/
 		// Check the certification key
-		if actual.GetCertification().Key != example.expectedCertKey {
-			t.Errorf("Expected: `%s`, Actual: `%s`", actual.GetCertification().Key, example.expectedCertKey)
+		if actual.GetCertification().GetKey() != example.expectedCertKey {
+			t.Errorf("Expected: `%s`, Actual: `%s`", actual.GetCertification().GetKey(), example.expectedCertKey)
 		}
 	}
 }
